fix player animation

+++ src/SpriteAnimation.java
@@ -11,17 +11,17 @@ import javafx.util.Duration;
 public class SpriteAnimation extends Transition {
     private final ImageView imageView;
     private int columns;
-    private int offsetX;
-    private int offsetY;
-    private int width;
-    private int height;
+    private double offsetX;
+    private double offsetY;
+    private double width;
+    private double height;
 
     public SpriteAnimation(
             ImageView imageView,
             Duration duration,
             int columns,
-            int offsetX, int offsetY,
-            int width, int height
+            double offsetX, double offsetY,
+            double width, double height
     ){
         this.imageView = imageView;
         this.columns = columns;
@@ -33,26 +33,25 @@ public class SpriteAnimation extends Transition {
         setCycleCount(Animation.INDEFINITE);
         setInterpolator(Interpolator.LINEAR);
         this.imageView.setViewport(new Rectangle2D(offsetX, offsetY, width, height));
-
     }
-    public void setOffsetX(int x){
+    public void setOffsetX(double x){
         this.offsetX = x;
     }
-    public void setOffsetY(int y){
+    public void setOffsetY(double y){
         this.offsetY = y;
     }
     public void setColumns(int columns){
         this.columns = columns;
     }
-    public void setWidthHeight(int width, int height){
+    public void setWidthHeight(double width, double height){
         this.width = width;
         this.height = height;
     }
 
     public void interpolate(double frac) { // frac = [0..1]
         final int index = Math.min((int) Math.floor(columns * frac), columns - 1);
-        final int x = (index % columns) * width + offsetX;
-        final int y = (index / columns) * height + offsetY;
+        final double x = (index % columns) * width + offsetX;
+        final double y = (index / columns) * height + offsetY;
         imageView.setViewport(new Rectangle2D(x, y, width, height));
     }
 }
